restructure manages system plugin.
