--- org.palladiosimulator.analyzer.slingshot.injection/src/org/palladiosimulator/analyzer/slingshot/injection/events/PlanEvent.java
@@ -1,13 +1,0 @@
-package org.palladiosimulator.analyzer.slingshot.injection.events;
-
-import org.palladiosimulator.analyzer.slingshot.injection.data.Plan;
-
-public interface PlanEvent {
-
-    /**
-     * TODO
-     *
-     * @return
-     */
-    public Plan getPlan();
-}
