+++ org.palladiosimulator.analyzer.slingshot.injection/src/org/palladiosimulator/analyzer/slingshot/managedSystem/injection/events/PlanEvent.java
@@ -1,0 +1,13 @@
+package org.palladiosimulator.analyzer.slingshot.managedSystem.injection.events;
+
+import org.palladiosimulator.analyzer.slingshot.managedSystem.injection.data.Plan;
+
+public interface PlanEvent {
+
+    /**
+     * TODO
+     *
+     * @return
+     */
+    public Plan getPlan();
+}
